docs(grid): add initial documentation for grid, split directives
references #6050 references #7508
