+++ src/index.ts
@@ -67,7 +67,9 @@ import { ClickBlock } from './util/click-block';
 import { Content } from './components/content/content';
 import { DateTime } from './components/datetime/datetime';
 import { FabContainer, FabButton, FabList } from './components/fab/fab';
-import { Grid, Row, Col } from './components/grid/grid';
+import { Col } from './components/grid/column';
+import { Grid } from './components/grid/grid';
+import { Row } from './components/grid/row';
 import { Icon } from './components/icon/icon';
 import { Img } from './components/img/img';
 import { InfiniteScroll } from './components/infinite-scroll/infinite-scroll';
@@ -138,7 +140,9 @@ export { ClickBlock } from './util/click-block';
 export { Content, ScrollEvent } from './components/content/content';
 export { DateTime } from './components/datetime/datetime';
 export { FabContainer, FabButton, FabList } from './components/fab/fab';
-export { Grid, Row, Col } from './components/grid/grid';
+export { Col } from './components/grid/column';
+export { Grid } from './components/grid/grid';
+export { Row } from './components/grid/row';
 export { Ion } from './components/ion';
 export { Icon } from './components/icon/icon';
 export { Img } from './components/img/img';
